ss_backend/requirements.txt
Django==3.2.7
phonenumbers==8.12.31
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
django-cors-headers==3.8.0
